pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
name = "feishu-token-mcp"
version = "1.0.0"
description = "A Model Context Protocol server for managing and automatically refreshing Feishu app and user access tokens"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.14.0",
    "smithery>=0.2.4",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
# Run the MCP server in development mode
dev = "smithery.cli.dev:main"
# Run server with interactive testing playground
playground = "smithery.cli.playground:main"

[tool.smithery]
# Function that returns a patched FastMCP server
server = "hello_server.server:create_server"